Revert "make sure we run python2"
According to jedbrown:
A lot of systems don't have a python2, just python and python2.? (includes Debian, RHEL, and Mac). The script is small that the best solution is probably to make it also
valid python3. At present, I think Arch Linux is the only distribution where /usr/bin/python points to python3 and Arch users are used to needing to run third-party scripts
with python2, so I would suggest reverting this change.

This reverts commit e5027ffbeb96446e7f6d95bbb73cd3daea06debb.